+++ compiler/gen/sql/testdata/golden/bidi_query_pkg.go
@@ -27,9 +27,7 @@ type PostQuery struct {
 	order      []func(*sql.Selector)
 	modifiers  []func(*sql.Selector)
 	inters     *entity.InterceptorStore
-	withFKs    bool
 	withAuthor *UserQuery
-	loadTotal  []func(context.Context, []*entity.Post) error
 	path       func(context.Context) (*sql.Selector, error)
 }
 
@@ -74,7 +72,6 @@ func (q *PostQuery) WithEdgeLoad(name string, opts ...runtime.LoadOption) runtim
 			q.withAuthor = NewUserQuery(q.config)
 			q.withAuthor.inters = q.inters
 		}
-		q.withFKs = true
 		q.withAuthor.applyLoad(runtime.NewLoadConfig(opts...), false)
 		return q.withAuthor
 	}
@@ -156,7 +153,6 @@ func (q *PostQuery) WithAuthor(opts ...func(entity.UserQuerier)) entity.PostQuer
 		opt(tq)
 	}
 	q.withAuthor = tq
-	q.withFKs = true
 	return q
 }
 
@@ -206,11 +202,6 @@ func (q *PostQuery) eagerLoad(ctx context.Context, nodes []*entity.Post) error {
 		}, func(n *entity.Post, e *entity.User) {
 			n.Edges.Author = e
 		}); err != nil {
-			return err
-		}
-	}
-	for i := range q.loadTotal {
-		if err := q.loadTotal[i](ctx, nodes); err != nil {
 			return err
 		}
 	}
@@ -587,7 +578,7 @@ func (q *PostQuery) GetModifiers() []func(*sql.Selector) {
 
 // GetWithFKs returns whether FK columns should be included. Implements runtime.QueryReader.
 func (q *PostQuery) GetWithFKs() bool {
-	return q.withFKs
+	return false
 }
 
 // Select allows the selection of one or more fields/columns for the given query,
@@ -768,13 +759,11 @@ func (q *PostQuery) clone() *PostQuery {
 		config:     q.config,
 		ctx:        q.ctx.Clone(),
 		inters:     q.inters,
-		loadTotal:  runtime.CloneSlice(q.loadTotal),
 		modifiers:  runtime.CloneSlice(q.modifiers),
 		order:      runtime.CloneSlice(q.order),
 		path:       q.path,
 		predicates: runtime.CloneSlice(q.predicates),
 		withAuthor: q.withAuthor.clone(),
-		withFKs:    q.withFKs,
 	}
 	return c
 }